+++ src/unitmodules.cpp
@@ -45,7 +45,7 @@ PUnitModule UFcn::Get_child(unsigned int n) const { return n==0?_next:nullptr; }
 void UFcn::connect(const PUnitModule m) { _next=m;_enable=true; }
 UFcn::UFcn(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _f = [](double u){return u;};
     _fu = nullptr;
     _next = nullptr;
@@ -77,7 +77,7 @@ void UFcnMISO::Module_Reset() {}
 int UFcnMISO::Get_childCnt() const { return _next.size(); }
 UFcnMISO::UFcnMISO(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _f = [](double *u){return u[0];};
     _fu = nullptr;
     UNITMODULE_INIT();
@@ -140,7 +140,7 @@ void UGain::connect(const PUnitModule m) { _next=m;_enable=true; }
 UGain::UGain(Simulator *sim, std::string name): UnitModule(sim, name)
 {
     _gain = 1;
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _next = nullptr;
     UNITMODULE_INIT();
 }
@@ -172,7 +172,7 @@ void UInput::Set_Continuous(bool isContinuous) { _isc=isContinuous; }
 void UInput::Set_SampleTime(double time) { _T=time; }
 UInput::UInput(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _cnt = -1;
     _isc = true;
     _f = [](double t){return 1.0;};
@@ -211,7 +211,7 @@ void UInput::Module_Update(double time)
 void UInput::Module_Reset()
 {
     _cnt = -1;
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
 }
 
 
@@ -258,7 +258,7 @@ void UNoise::Set_SampleTime(double time) { _T=time;_ltn=-_T; }
 UNoise::UNoise(Simulator *sim, std::string name): UnitModule(sim, name)
 {
     DISCRETE_INITIALIZE(-1);
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     UNITMODULE_INIT();
     _enable = true;
     _mean = 0;
@@ -327,7 +327,7 @@ void UProduct::Module_Reset() {}
 int UProduct::Get_childCnt() const { return _next.size(); }
 UProduct::UProduct(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     UNITMODULE_INIT();
 }
 void UProduct::Set_InputGain(double inputgain, int port)
@@ -396,7 +396,7 @@ int USum::Get_childCnt() const { return _next.size(); }
 void USum::Set_Redundant(bool rdnt) { _rdnt=rdnt; };
 USum::USum(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _rdnt = true;
     UNITMODULE_INIT();
 }
@@ -469,7 +469,7 @@ PUnitModule UTransportDelay::Get_child(unsigned int n) const { return n==0?_next
 void UTransportDelay::connect(const PUnitModule m) { _next=m;_enable=true; }
 UTransportDelay::UTransportDelay(Simulator *sim, std::string name): UnitModule(sim, name)
 {
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _iv = 0;
     _lv.push_back(0);
     _next = nullptr;
@@ -566,7 +566,7 @@ void UZOH::Set_SampleTime(double time) { _T=time;_ltn=-_T; }
 UZOH::UZOH(Simulator *sim, std::string name): UnitModule(sim, name)
 {
     DISCRETE_INITIALIZE(1);
-    _outvalue = nan("");
+    _outvalue = 0.0/0.0;
     _next = nullptr;
     UNITMODULE_INIT();
 }