+++ changelog.md
@@ -53,3 +53,4 @@
 
 # V2.0.11
 - [CMakeLists.txt] 将依赖库`tracelog`由必装改为选装.
+- [CMakeLists.txt] CHANGED: `nan("")` to `0.0/0.0`.